Update index.html
I should have added in the descriptions the changes I made as I progressed through this lab, but this commit is to comment on the changes I made overall. I organized the section of my app and decided to be a bit creative with my paragraphs and listings. I added an unordered (bullet) list of my personal hobbies. I also made a few other additions to align with the instructions for this assignment